% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{getLinkFromPage}
\alias{getLinkFromPage}
\title{RCSB Link Scraper Function}
\usage{
getLinkFromPage(url, path, prepend = "http://www.rcsb.org")
}
\arguments{
\item{url}{The url corresponding to the RCSB Structure Page in question}

\item{path}{The path of the <a> element you're wanting to scrape a link from}

\item{prepend}{Any string you want to prepend to the results}
}
\description{
This function allows you to scrape a unique link specified by an XPath,
applied on an individual RCSB structure page. Concretely, this lets you
obtain links to assets (like the FASTA file, PDB file, mmCIF file, etc.)
if you have the corresponding XPath for that <a> element.
}
\details{
To apply this to a vector, see `getLinksFromPages``
}
\examples{
# This gets the FASTA file link associated with the 6B4V structure on RCSB
url <- "http://www.rcsb.org/structure/6B4V"
fasta_path <- '//*[@id="DownloadFilesButton"]/ul/li[1]/a'
getLinkFromPage(url, fasta_path)
}
\keyword{PDB}
\keyword{RSCB}
\keyword{structure}
