% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_functions.R
\name{getPDBResults}
\alias{getPDBResults}
\title{RCSB Search Result Scraper Function}
\usage{
getPDBResults(path = "data/searchResults.html")
}
\arguments{
\item{path}{The path to the HTML search file}
}
\description{
This function allows you to parse (to df) the basic results of a search on
RCSB, based on attribute XPaths derived by Elliot Williams in February 2018.
}
\details{
You want to save the HTML associated with a given search (with all search
results displayed) to disk before running this command.
}
\examples{
getPDBResults(path="data/searchResults.html")
}
\keyword{PDB}
\keyword{RSCB}
